Utilizado coroutine en vez de Thread a la hora de ejecutar la base de datos
