+++ app/build.gradle
@@ -42,6 +42,10 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
+    //Coroutine
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0'
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.0"
+
     //Room library
     def room_version = "2.3.0"
     implementation "androidx.room:room-runtime:$room_version"
