+++ app/src/main/java/com/example/aad_playground/ut03/ex02/presentation/Example02Activity.kt
@@ -2,7 +2,6 @@ package com.example.aad_playground.ut03.ex02.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import com.example.aad_playground.R
 import com.example.aad_playground.ut03.ex02.data.PersonDataRepository
 import com.example.aad_playground.ut03.ex02.data.PersonLocalSource
@@ -16,26 +15,16 @@ class Example02Activity : AppCompatActivity() {
         PersonDataRepository(PersonLocalSource(applicationContext))
     }
 
+    private lateinit var viewModel: Example02ViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_example02)
+        viewModel = Example02ViewModel()
         executeQuery()
     }
 
     private fun executeQuery() {
-        Thread {
-            val carList: MutableList<CarModel> = mutableListOf(
-                CarModel(1, "Audi", "A3"),
-                CarModel(2, "Audi", "A5")
-            )
-            val jobList: MutableList<JobModel> = mutableListOf(
-                JobModel(1, "Programador"),
-                JobModel(2, "JefeProyecto")
-            )
-            repository.savePerson(PersonModel(1, "Name01", 1, "1", PetModel(1, "Pet01", 2), carList, jobList))
-            val people = repository.findPersonAndPetAndCarAndJob()
-            Log.d(TAG, "$people")
-        }.start()
+        viewModel.getPeople(repository)
     }
 }
